requirements.txt
# Requirements for ML Model Optimization Environment
# Install these packages in the Docker container

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Quantization and optimization
bitsandbytes>=0.41.0
optimum>=1.12.0
auto-gptq>=0.4.0
peft>=0.4.0
quanto>=0.1.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0
speechbrain>=0.5.0
whisper-openai>=20231117

# Computer vision and image processing
opencv-python>=4.8.0
pillow>=9.5.0
imageio>=2.31.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Sentence transformers and embeddings
sentence-transformers>=2.2.0

# Model serving and optimization
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
tensorrt>=8.6.0
triton>=2.1.0

# Monitoring and visualization
tensorboard>=2.13.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and development
jupyter>=1.0.0
ipywidgets>=8.0.0
jupyterlab>=4.0.0

# System monitoring
psutil>=5.9.0
nvidia-ml-py>=12.0.0
GPUtil>=1.4.0

# Data handling
datasets>=2.14.0
evaluate>=0.4.0
tqdm>=4.65.0

# Utilities
requests>=2.31.0
boto3>=1.28.0
pyyaml>=6.0.0
click>=8.1.0
